+++ packages/destination-actions/src/destinations/amplitude/logEventV2/autocapture-fields.ts
